Added temporary dataset for commodity price
As we can not quickly identify a way to convert the .xlsb format to
a readily usable format, we simply copied the data from the .xlsb
and saved it as 'igc_goi.csv'. Then the R script performs some basic
manipulation before saving it in the .feather format.

We will later automate the whole data pipeline.